结果不对，没找到很好的方式debug

+++ func.h
@@ -15,13 +15,13 @@
 using namespace std;
 
 
-void secretShareLL(long long data, long long &s1, long long &s2);
+void secretShareLL(cpp_int data, cpp_int &s1, cpp_int &s2);
 
-void secretShareInt(long long data, int &d1, int &d2);
+void secretShareInt(cpp_int data, int &d1, int &d2);
 
-vector<vector<long long>> readData(string filename);
+vector<vector<cpp_int>> readData(string filename);
 
-vector<point*> data2Point(vector<vector<long long>>&dataList, vector<point*>&pList1, vector<point*>&pList2);
+vector<point*> data2Point(vector<vector<cpp_int>>&dataList, vector<point*>&pList1, vector<point*>&pList2);
 
 void generateBeaverSet(vector<vector<int>>&set1, vector<vector<int>>&set2);
